CRUD of salary and menu of manager

+++ EmployeeDatabase/include/DBmanage.h
@@ -21,7 +21,7 @@ public:
 	sqlite3_stmt* stmt {};
 	bool open(const char* str);
 	bool close();
-	int executeQuery(const char* sql , int count = 0);
+	int executeQuery(const char* sql , float count = 0);
 	bool selectQuery(const char* sql);
 	static Database& getInstance() {
 		static Database db;
